// src/app/global/global-interface.ts
export enum Season { regularSeason, postSeason }//TODO - regularSeason, postSeason
export enum Conference { AFC, NFC, ACC, AmericanAthletic, BigTen, ConferenceUSA, MidAmerican, MountainWest, Pac12, SEC, Big12, IAIndependents, SunBelt} //to get string value: Conference[myValue], where myValues is a Conference value

export enum Division { AFCEast, AFCNorth, AFCSouth, AFCWest, NFCEast, NFCNorth, NFCSouth, NFCWest, ATLANTIC, COASTAL, EAST, WEST, East, West, MOUNTAIN, NORTH, SOUTH, Big12, IAIndependents, SunBelt }  //to get string value: Year[myValue], where myValues is a Year value

export interface SportPageParameters {
  conference?: Conference;
  division?: Division;
  year?: number;
  season?: Season;
  teamId?: number;
  playerId? :number;
  type?: string;
  scope?: string;
}
// ...
export interface HeadlineData {
    event: number;
    displayHeadline: string;
    metaHeadline: string;
    dateline: string;
    article: string;
    commentHeader: string;
    leftColumn: {
        "about-the-teams": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "historical-team-statistics": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "last-matchup": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "starting-lineup-home": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "starting-lineup-away": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "injuries-home": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "injuries-away": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "upcoming-game": {
            displayHeadline: string;
            metaHeadline: string;
        };
    };
    rightColumn: {
        "pitcher-player-comparison": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "catcher-player-comparison": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "first-base-player-comparison": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "second-base-player-comparison": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "third-base-player-comparison": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "shortstop-player-comparison": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "left-field-player-comparison": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "center-field-player-comparison": {
            displayHeadline: string;
            metaHeadline: string;
        };
        "right-field-player-comparison": {
            displayHeadline: string;
            metaHeadline: string;
        };
    };
    home: {
        id: number;
        location: string;
        name: string;
        hex: string;
        logo: string;
        images: string;
        wins: number;
        losses: number;
    }
    away: {
        id: number;
        location: string;
        name: string;
        hex: string;
        logo: string;
        images: string;
        wins: number;
        losses: number;
    }
}